Refactor syscall file_stream
